# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orthoplex API",
        "description": "A Laravel 12 multi-tenant SaaS API with comprehensive authentication, RBAC, and user management.",
        "contact": {
            "email": "support@orthoplex.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://orthoplex.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List users with advanced querying",
                "description": "Get paginated list of users with RSQL filtering, cursor pagination, and advanced sorting",
                "operationId": "d19efc9968cc35819cb1c9b33a094ed2",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "RSQL filter expression",
                        "schema": {
                            "type": "string",
                            "example": "name=like=john;email!=admin@example.com"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Global text search across name and email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort fields (comma-separated, prefix with - for desc)",
                        "schema": {
                            "type": "string",
                            "example": "name,-created_at"
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Select specific fields to return",
                        "schema": {
                            "type": "string",
                            "example": "id,name,email,created_at"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for pagination",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Pagination direction",
                        "schema": {
                            "type": "string",
                            "default": "next",
                            "enum": [
                                "next",
                                "prev"
                            ]
                        }
                    },
                    {
                        "name": "include_total",
                        "in": "query",
                        "description": "Include total count (expensive operation)",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by creation date start",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by creation date end",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "cursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next_cursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev_cursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "has_more": {
                                                    "type": "boolean"
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Create a new user account",
                "operationId": "867bb38a7e18614dbc341116c2c5ff01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "UTC"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "member"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Soft delete a user (can be restored)",
                "operationId": "e316536e997354eb25c630e84d7169c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}/restore": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Restore deleted user",
                "description": "Restore a soft-deleted user",
                "operationId": "c18e3460ac1996adcfe5a87c5fedfe97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User restored successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/users/top-logins": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users with most logins",
                "description": "Get users ordered by login count within time window",
                "operationId": "43d1ed8c2f20215b7c7793f8f3d3e5dd",
                "parameters": [
                    {
                        "name": "window",
                        "in": "query",
                        "description": "Time window",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "7d",
                                "30d"
                            ],
                            "example": "7d"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of users to return",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top login users retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}